requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
